Rename to FAIRDataPoint-index (#1)
