+++ Dockerfile
@@ -2,6 +2,6 @@ FROM openjdk:11-jdk-slim
 
 WORKDIR /app
 
-ADD target/fair-metadata-index-0.1.0-SNAPSHOT.jar /app/app.jar
+ADD target/fairdatapoint-index.jar /app/app.jar
 
 ENTRYPOINT java -jar app.jar
