+++ src/main/java/solutions/fairdata/fdp/index/domain/package-info.java
@@ -20,27 +20,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package solutions.fairdata.metadata.index.storage;
 
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
-
-@Configuration
-@EnableRedisRepositories
-public class StorageConfig {
-    @Bean
-    public RedisConnectionFactory connectionFactory() {
-        return new LettuceConnectionFactory();
-    }
-    
-    @Bean
-    public RedisTemplate<?, ?> redisTemplate() {
-        var template = new RedisTemplate<byte[], byte[]>();
-        template.setConnectionFactory(connectionFactory());
-        return template;
-    }
-}
+@javax.annotation.ParametersAreNonnullByDefault
+package solutions.fairdata.fdp.index.domain;
