+++ src/main/java/solutions/fairdata/fdp/index/web/WebConfig.java
@@ -20,10 +20,16 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package solutions.fairdata.metadata.index.storage;
+package solutions.fairdata.fdp.index.web;
 
-import org.springframework.data.repository.CrudRepository;
-import solutions.fairdata.metadata.index.domain.IndexEntry;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import solutions.fairdata.fdp.index.service.ServiceConfig;
 
-public interface EntryRepository extends CrudRepository<IndexEntry, String> {
+@Configuration
+@ComponentScan
+@Import(ServiceConfig.class)
+public class WebConfig implements WebMvcConfigurer {
 }
