+++ src/main/java/solutions/fairdata/fdp/index/web/controller/package-info.java
@@ -20,16 +20,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package solutions.fairdata.metadata.index.web;
 
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import solutions.fairdata.metadata.index.service.ServiceConfig;
-
-@Configuration
-@ComponentScan
-@Import(ServiceConfig.class)
-public class WebConfig implements WebMvcConfigurer {
-}
+@javax.annotation.ParametersAreNonnullByDefault
+package solutions.fairdata.fdp.index.web.controller;
